SupermercatoBootrap_PaginaProdotto_daCompletare

+++ 04 - Assignment/19_SupermercatoAdvanced/Models/Cliente.cs
@@ -7,5 +7,5 @@ public class Cliente
     public List<Prodotto>  Carrello { get; set; }
     public List<Prodotto> StoricoAcquisti { get; set; }
     public int PercentualeSconto { get; set; }
-    public double Credito { get; set; }
+    public decimal Credito { get; set; }
 }
